ADD: CategoriesContainer test 완성
