+++ src/services/__mock__/api.js
@@ -1,0 +1,7 @@
+export async function regionData() {
+  return [];
+}
+
+export async function categoryData() {
+  return [];
+}
