--- src/services/__mock__/fetchData.js
@@ -1,3 +1,0 @@
-export async function regionData() {
-  return [];
-}
